+++ core/AppSpec.js
@@ -7,6 +7,7 @@
 import {XH} from '@xh/hoist/core';
 import {throwIf} from '@xh/hoist/utils/js';
 import {isFunction, isNil, isString} from 'lodash';
+import {apiRemoved} from '../utils/js';
 
 /**
  * Object used to hold the specification for a client-side Hoist application.
@@ -39,10 +40,10 @@ export class AppSpec {
      *      app has loaded and fully initialized, including elapsed time of asset loading and init.
      * @param {boolean} [c.webSocketsEnabled] - true to enable Hoist websocket connectivity,
      *      establish a connection and initiate a heartbeat..
-     * @param {boolean} [c.idleDetectionEnabled] - true to enable auto-suspension by `IdleService`.
-     * @param {Class} [c.idleDialogClass] - Component or element factory used to indicate App has
+     * @param {(Class|function)} [c.idlePanel] - Optional custom Component to display when App has
      *      been suspended.  The component will receive a single prop -- onReactivate -- a callback
-     *      called when user has acknowledged the suspension and wishes to reload the app and continue working.
+     *      called when the user has acknowledged the suspension and wishes to reload the app and
+     *      continue working.  Specify as a React Component or an element factory.
      * @param {?string} [c.loginMessage] - Optional message to show on login form (for non-SSO apps).
      * @param {?string} [c.lockoutMessage] - Optional message to show users when denied access to app.
      */
@@ -57,10 +58,10 @@ export class AppSpec {
         checkAccess,
         trackAppLoad = true,
         webSocketsEnabled = false,
-        idleDetectionEnabled = false,
-        idleDialogClass = null,
+        idlePanel = null,
         loginMessage = null,
-        lockoutMessage = null
+        lockoutMessage = null,
+        ...rest
     }) {
         throwIf(!modelClass, 'A Hoist App must define a modelClass.');
         throwIf(!componentClass, 'A Hoist App must define a componentClass');
@@ -70,6 +71,7 @@ export class AppSpec {
             `Please import and provide containerClass from "@xh/hoist/${isMobileApp ? 'mobile' : 'desktop'}/AppContainer".`
         );
         throwIf(isNil(isSSO), 'A Hoist App must define isSSO');
+        apiRemoved(rest.idleDetectionEnabled, 'idleDetectionEnabled', 'Set "xhIdleConfig" in configuration instead.');
 
         throwIf(
             !isString(checkAccess) && !isFunction(checkAccess),
@@ -88,8 +90,7 @@ export class AppSpec {
         this.trackAppLoad = trackAppLoad;
 
         this.webSocketsEnabled = webSocketsEnabled;
-        this.idleDetectionEnabled = idleDetectionEnabled;
-        this.idleDialogClass = idleDialogClass;
+        this.idlePanel = idlePanel;
         this.loginMessage = loginMessage;
         this.lockoutMessage = lockoutMessage;
         Object.freeze(this);